+++ examples/polygon.py
@@ -1,5 +1,4 @@
 import numpy as np
-
 from pypex.poly2d import polygon
 
 
@@ -7,7 +6,7 @@ def main():
     poly1 = polygon.Polygon([[0.0, 0.0], [1.0, 0.0], [0.0, 1.0], [1.0, 1.0]])
     poly2 = polygon.Polygon([[0.5, 0.3], [0.0, -1.0], [1.0, -1.0]])
 
-    print("Polygon with hull defined by {} is automaticaly sorted to clokwise corners as {}\n"
+    print("Polygon with hull defined by {} \n is automaticaly sorted to clokwise corners as {}\n"
           "".format([[0.0, 0.0], [1.0, 0.0], [0.0, 1.0], [1.0, 1.0]], poly1.hull))
 
     try:
@@ -16,7 +15,11 @@ def main():
         from matplotlib.patches import Polygon as polyg
         from matplotlib.collections import PatchCollection
 
-        ig, ax = plt.subplots()
+        # mpl rcParams
+        params = {'legend.fontsize': 8, 'legend.handlelength': 0.5, "font.size": 8}
+        matplotlib.rcParams.update(params)
+
+        fig, ax = plt.subplots()
         patches = [polyg(poly1.hull, True), polyg(poly2.hull, True)]
         p = PatchCollection(patches, cmap=matplotlib.cm.jet, alpha=0.4)
         colors = 100 * np.random.rand(len(patches))
@@ -46,16 +49,48 @@ def main():
         from matplotlib.patches import Polygon as polyg
         from matplotlib.collections import PatchCollection
 
-        ig, ax = plt.subplots()
+        # mpl rcParams
+        params = {'legend.fontsize': 8, 'legend.handlelength': 0.5, "font.size": 8}
+        matplotlib.rcParams.update(params)
+
+        fig, ax = plt.subplots()
         patches = [polyg(poly1.hull, True), polyg(poly2.hull, True)]
-        p = PatchCollection(patches, cmap=matplotlib.cm.jet, alpha=0.4)
+        p = PatchCollection(patches, cmap=matplotlib.cm.jet, alpha=0.4, edgecolors="k")
         colors = 100 * np.random.rand(len(patches))
         p.set_array(np.array(colors))
         ax.add_collection(p)
         plt.scatter(intersection.hull.T[0], intersection.hull.T[1])
-        plt.xlim(-2, 2)
-        plt.ylim(-2, 2)
+        plt.xlim(-0.5, 1.5)
+        plt.ylim(-1.5, 1.5)
         plt.show()
+    except ImportError:
+        pass
+
+    # inpolygon
+    _polygon = np.array([[0.0, 0.0], [0.3, 0.0], [0.4, 1.1], [0.1, 0.5]])
+    poly = polygon.Polygon(_polygon)
+    inpolygon = poly.inpolygon()
+
+    try:
+        import matplotlib
+        import matplotlib.pyplot as plt
+        from matplotlib.patches import Polygon as polyg
+        from matplotlib.collections import PatchCollection
+
+        # mpl rcParams
+        params = {'legend.fontsize': 8, 'legend.handlelength': 0.5, "font.size": 8}
+        matplotlib.rcParams.update(params)
+
+        fig, ax = plt.subplots()
+        patches = [polyg(poly.hull, True), polyg(inpolygon.hull, True)]
+        p = PatchCollection(patches, cmap=matplotlib.cm.jet, alpha=0.4, edgecolors="k")
+        colors = 100 * np.random.rand(len(patches))
+        p.set_array(np.array(colors))
+        ax.add_collection(p)
+        plt.xlim(-0.25, 0.6)
+        plt.ylim(-0.25, 1.2)
+        plt.show()
+
     except ImportError:
         pass