+++ src/pypex/__init__.py
@@ -1,0 +1,5 @@
+__version__ = "0.2.0"
+
+from pypex.poly2d.polygon import Point, Polygon
+from pypex.poly2d.line import Line
+from pypex.poly2d import projection