+++ src/pypex/poly2d/line.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 from pypex.base import shape
 from pypex.poly2d.intersection import linter
 from pypex.base.conf import ROUND_PRECISION
@@ -61,5 +63,28 @@ class Line(shape.Shape2D):
             return
         return intersection[2]
 
+    def to_array(self):
+        """
+        Get points of line in numpy array.
+        :return: numpy.array;
+        """
+        return np.array([point.to_array() for point in self.to_Points()])
+
     def sort_clockwise(self, *args, **kwargs):
         return self.hull
+
+    def direction_vector(self):
+        """
+        Get direction vector.
+        :return: numpy.array;
+        """
+        return self.hull[1] - self.hull[0]
+
+    def parametrized(self):
+        """
+        Return callable parametrization of given line as function.
+        """
+        def _parametrized(t):
+            v = self.direction_vector()
+            return self.hull[0][0] + (t * v[0]), self.hull[0][1] + (t * v[1])
+        return _parametrized