+++ CHANGELOG.md
@@ -1,14 +1,27 @@
 # Change Log
 
-## [v0.1]()
+## [v0.1.0]()
 
-### Enhancements
+### Features
 ##### all in 2D
 
     - implementation of Polygon, Line and Point 
     - implementation of SAT (separating axis theroem) algorithm
     - implemnetaion of line intersection 
     - implementation of polygon intersection
+
+## [v0.2.0]()
+
+**Release date:** 2020-11-01
+
+### Features
+    
+    - inerpolygon - find polygon which points are defind as center of each edge of original polygon
+    - line parametrization method
+
+### Enhancements
+    
+    - support transpose of Shape2D (.T) for simplier interaction with raw points
 
 ### Fixes